style(git-hunk): decorations (#8)
* draft

* style(hunk): line diff background

* feat(hunk/decorations): include emphasized diffs

* fix: clippy warnings

* fix(info): decorations not overridden

* feat(hunk/inline-diff): use unicode grapheme tokenization

This makes the diffs looks better and finer. The previous version
splits the words by space, which results in coarser diff.

* fix(test_app/repo_git_hunks): failing tests

+++ src/position.rs
@@ -37,10 +37,10 @@ impl Position {
         }
     }
 
-    pub fn move_left(&self, by: i32) -> Position {
+    pub fn move_left(&self, by: usize) -> Position {
         Position {
             line: self.line,
-            column: self.column.saturating_sub(by as usize),
+            column: self.column.saturating_sub(by),
         }
     }
